Refactor increment delegate, update message event logic

+++ src/main/java/org/folio/rest/delegate/CrIncrementCheckCountDelegate.java
@@ -10,7 +10,7 @@ public class CrIncrementCheckCountDelegate extends AbstractLoggingDelegate {
   public void execute(DelegateExecution execution) throws Exception {
     log.info("Executing Increment Check Count Delegate");
 
-    Long newCount = ((Long) execution.getVariable("checkedCount") + 1);
+    Integer newCount = ((Integer) execution.getVariable("checkedCount") + 1);
 
     execution.setVariable("checkedCount", newCount);
     log.info("New count is {}", newCount);
